Refactor deploy code
