+++ deploy/python/infer_dataset.py
@@ -15,25 +15,18 @@
 import argparse
 import codecs
 import os
-import sys
 
 import time
 import yaml
 import numpy as np
 import paddle
 import paddle.nn.functional as F
-
-from paddle.inference import create_predictor, PrecisionType
+from paddle.inference import create_predictor
 from paddle.inference import Config as PredictConfig
 
-LOCAL_PATH = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(os.path.join(LOCAL_PATH, '..', '..'))
-
-import paddleseg
+import paddleseg.deploy.infer as infer_api
 from paddleseg.cvlibs import manager
 from paddleseg.utils import logger, metrics, progbar
-
-from infer import Predictor, DeployConfig, use_auto_tune
 
 
 def parse_args():
@@ -148,7 +141,7 @@ def get_dataset(args):
     if args.dataset_type not in comp.components_dict:
         raise RuntimeError("The dataset is not supported.")
 
-    cfg = DeployConfig(args.cfg)
+    cfg = infer_api.DeployConfig(args.cfg)
 
     if args.resize_width == 0 and args.resize_height == 0:
         transforms = cfg.transforms.transforms
@@ -162,7 +155,8 @@ def get_dataset(args):
             "type": "Resize",
             'target_size': [args.resize_width, args.resize_height]
         })
-        transforms = DeployConfig.load_transforms(transforms_dic).transforms
+        transforms = infer_api.DeployConfig.load_transforms(
+            transforms_dic).transforms
 
     kwargs = {
         'transforms': transforms,
@@ -192,11 +186,11 @@ def auto_tune(args, dataset, img_nums):
     """
     logger.info("Auto tune the dynamic shape for GPU TRT.")
 
-    assert use_auto_tune(args)
+    assert infer_api.use_auto_tune(args)
 
     num = min(len(dataset), img_nums)
 
-    cfg = DeployConfig(args.cfg)
+    cfg = infer_api.DeployConfig(args.cfg)
     pred_cfg = PredictConfig(cfg.model, cfg.params)
     pred_cfg.enable_use_gpu(100, 0)
     if not args.print_detail:
@@ -229,7 +223,7 @@ def auto_tune(args, dataset, img_nums):
     logger.info("Auto tune success.\n")
 
 
-class DatasetPredictor(Predictor):
+class DatasetPredictor(infer_api.Predictor):
     def __init__(self, args):
         super().__init__(args)
 
@@ -299,7 +293,7 @@ class DatasetPredictor(Predictor):
 
 
 def main(args):
-    if use_auto_tune(args):
+    if infer_api.use_auto_tune(args):
         dataset = get_dataset(args)
         tune_img_nums = 10
         auto_tune(args, dataset, tune_img_nums)
@@ -307,7 +301,7 @@ def main(args):
     predictor = DatasetPredictor(args)
     predictor.run_dataset()
 
-    if use_auto_tune(args) and \
+    if infer_api.use_auto_tune(args) and \
         os.path.exists(args.auto_tuned_shape_file):
         os.remove(args.auto_tuned_shape_file)
