+++ tools/data/cityscapes_trainid2labelid.py
@@ -20,15 +20,12 @@ Obtain the mIoU of Cityscapes test set:
 
 import argparse
 import os
-import sys
 from collections import namedtuple
 
 import numpy as np
 from PIL import Image
 from tqdm import tqdm
 
-LOCAL_PATH = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(os.path.join(LOCAL_PATH, '..'))
 from paddleseg.utils import get_image_list
